fix reportes.csv, iconos.png, fix visual celdas tipo zebra, image loader desde config, se agrega impresora

+++ demo/src/main/java/com/example/demo/dto/ProductoMasVendidoDto.java
@@ -24,7 +24,7 @@ public class ProductoMasVendidoDto {
         this.producto = producto;
     }
 
-    public double getCantidadVendida() {
+    public long getCantidadVendida() {
         return cantidadVendida;
     }
